# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Workout Tracker API",
    "version": "1.0.0",
    "description": "API documentation for the Workout Tracker application"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local Development Server"
    }
  ],
  "paths": {
...
    "/workout_logs": {
      "post": {
        "summary": "Create a new workout log",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkoutLog"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Workout log created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisplayWorkoutLog"
                }
              }
            }
          },
          "404": {
            "description": "Workout plan not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "get": {
        "summary": "Get all workout logs for the current user",
        "responses": {
          "200": {
            "description": "Successful retrieval of workout logs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DisplayWorkoutLog"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No workout log found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/workout_logs/{workout_log_id}/exercises": {
      "post": {
        "summary": "Add an exercise to a workout log",
        "parameters": [
          {
            "name": "workout_log_id",
            "in": "path",
            "description": "ID of the workout log to add exercise to",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AddExerciseToWorkoutLog"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Workout log exercise created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DisplayWorkoutLogExercise"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Workout log not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "get": {
        "summary": "Get exercises logged for a specific workout log",
        "parameters": [
          {
            "name": "workout_log_id",
            "in": "path",
            "description": "ID of the workout log to retrieve exercises for",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of workout log exercises",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DisplayWorkoutLogExercise"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No exercises logged for this workout"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/workouts": {
      "get": {
        "summary": "Get all workout plans",
        "responses": {
          "200": {
            "description": "Successful retrieval of workout plans",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DisplayWorkoutPlan"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No workout plan found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "Create a new workout plan",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkoutPlan"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Workout plan created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisplayWorkoutPlan"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/workouts/{plan_id}": {
      "get": {
        "summary": "Get workout plan by ID",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "description": "ID of the workout plan to retrieve",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of workout plan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisplayWorkoutPlan"
                }
              }
            }
          },
          "404": {
            "description": "Workout plan not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "put": {
        "summary": "Update a workout plan",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "description": "ID of the workout plan to update",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkoutPlan"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Workout plan updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisplayWorkoutPlan"
                }
              }
            }
          },
          "404": {
            "description": "Workout plan not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Delete a workout plan",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "description": "ID of the workout plan to delete",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Workout plan deleted successfully"
          },
          "404": {
            "description": "Workout plan not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateWorkoutLog": {
        "type": "object",
        "properties": {
          "workout_plan_id": {
            "type": "integer"
          },
          "duration": {
            "type": "integer"
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "required": ["workout_plan_id", "duration"]
      },
      "DisplayWorkoutLog": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer"
          },
          "workout_plan_id": {
            "type": "integer"
          },
          "duration": {
            "type": "integer"
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "required": ["date", "id", "workout_plan_id", "duration"]
      },
      "AddExerciseToWorkoutLog": {
        "type": "object",
        "properties": {
          "exercise_id": {
            "type": "integer"
          },
          "sets_completed": {
            "type": "integer"
          },
          "reps_completed": {
            "type": "integer"
          },
          "weight_used": {
            "type": "number",
            "format": "float",
            "nullable": true
          }
        },
        "required": ["exercise_id", "sets_completed", "reps_completed"]
      },
      "DisplayWorkoutLogExercise": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "workout_log_id": {
            "type": "integer"
          },
          "exercise_id": {
            "type": "integer"
          },
          "sets_completed": {
            "type": "integer"
          },
          "reps_completed": {
            "type": "integer"
          },
          "weight_used": {
            "type": "number",
            "format": "float",
            "nullable": true
          }
        },
        "required": ["id", "workout_log_id", "exercise_id", "sets_completed", "reps_completed"]
      },
      "CreateWorkoutPlan": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "weeks": {
            "type": "integer",
            "nullable": true
          }
        },
        "required": ["name"]
      },
      "DisplayWorkoutPlan": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "user_id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["id", "user_id", "name", "created_at", "updated_at"]
      },
      "UpdateWorkoutPlan": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}